commit
by running this file you can automate your process of running all files one by one. you can set the sequence of running the files and when you run this file it automatically run all the files one by one according to sequence. And also in this file we can create our secret in which we can give our database password, so this is good for security purpose. because only one who run this file know the password of database.